strings: Move str to int conversion functions to strings module
